requirements.txt
# Core
pydantic>=2.7.0
python-dotenv>=1.0.0

# PDF Processing
PyPDF2>=3.0.0

# Web Scraping
beautifulsoup4>=4.12.0
requests>=2.31.0
lxml>=5.0.0

# Reddit API
praw>=7.7.0

# OpenAI
openai>=1.40.0

# Async
aiohttp>=3.9.0

# ML/NLP
numpy>=1.26.0
scikit-learn>=1.4.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-mock>=3.12.0

# Utilities
tqdm>=4.66.0

# API Server
flask>=3.0.0
flask-cors>=4.0.0
